+++ marimo/_runtime/packages/module_registry.py
@@ -1,3 +1,6 @@
+# Copyright 2024 Marimo. All rights reserved.
+from __future__ import annotations
+
 import importlib.util
 
 from marimo._ast.cell import CellId_t
@@ -6,11 +9,13 @@ from marimo._runtime.dataflow import DirectedGraph
 
 class ModuleRegistry:
     def __init__(
-        self, graph: DirectedGraph, excluded_modules: set[str]
+        self, graph: DirectedGraph, excluded_modules: set[str] | None = None
     ) -> None:
         self.graph = graph
         # modules that do not have corresponding packages on package index
-        self.excluded_modules = excluded_modules
+        self.excluded_modules = (
+            excluded_modules if excluded_modules is not None else set()
+        )
 
     def defining_cell(self, module_name: str) -> CellId_t | None:
         """Get the cell id of the cell importing module_name"""
