+++ marimo/_runtime/packages/utils.py
@@ -1,3 +1,4 @@
+# Copyright 2024 Marimo. All rights reserved.
 import os
 import sys
